Node.js , express, pug, mongodb
Look in this path nodejs>basic>myapp

you must be running the npm start command and the mongod command for the server to be working and for the db to be running. I'm currently having connection issues which will be solved soon. But everything is set up for the most part. Also I have the web pages now running using the pug template which corresponds with ecpress and node.js .