updated map.html
now has button to open submission form that posts building name and busy level (not in array yet) . Next step is to get a schema created that makes busy level an array . After that will be to to be able to update the already existing entrees in the registrations. After that will be to have a form that is pushed via buttons with inserted building name from button instead of user text submitted button names.