Updated Submission Form
Updated to now push time , it is calculated by (hours * 60 + minutes) also The busy level buttons now automatically submit when pressed. The name is automatically added to the form for submission, the next step is to pull the name from a button, I need the code that Daniel and Rolando worked on to get the names automatically from buttons.

After that I will be working on getting info pulled as well updating existing entrees rather than creating a new one with every submission.